% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the integer number representing each state}

\item{year}{integers years from 2013 to 2015 inclusive}
}
\value{
latitude and longitude of the state
If the state.num is not valid, it will thrown an error
if the longtitude exceeds 900 it will throw an error
if the latitue exceeds 90 it will throw and error
it will tell user if there are no accidents meeting criteria
a chart will be produced if input is good
}
\description{
fars_map_state
}
\examples{
fars_map_state(1,2015)
}
