% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name and directory of input csv file}
}
\value{
returns data from the file as a data frame
}
\description{
fars_read
}
\examples{
fars_read( system.file("extdata","accident_2015.csv.bz2",package="BuildRWeek2Assign0207"))
}
