% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of integer years between 2013 and 2015 inclusive}
}
\value{
returns the years and months found in the selected datafile
}
\description{
fars_read_years
}
\examples{
 fars_read_years(c(2013,2014))
}
