% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years
will count entries by year aross month for given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{as a list of integers years from 2013 to 2015 inclusive}
}
\value{
returns total entries (accidents) in data across year by month
}
\description{
fars_summarize_years
will count entries by year aross month for given year
}
\examples{
 fars_summarize_years(c(2013,2014,2015))
}
