% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{-integer value of year from 2013 to 2015 inclusive}
}
\value{
returns the full file name (wihtout the directory)
}
\description{
will recreate the file name associated wtih the given integer year
}
\examples{
make_filename(2015)
}
